requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
wordcloud>=1.9.0
scikit-learn>=1.2.0
nltk>=3.8.0
tensorboard>=2.13.0
pillow>=9.5.0
tqdm>=4.65.0
kaggle>=1.5.13 